test: add backbone testing with testscripts
This change is a port from the work I did in `gnols` (ancestor of this
repo) to have a proper testing framework.

It uses testscripts to setup a comprehensive and extensible way to test
the gnopls server. The setup consists of running the server and
mimicking a LSP client via the "lsp" command used in the txtar files.

Each txtar file is used to assert a specific command or functionnality:
- testdata/initialize.txtar asserts the lsp initialize command
- testdata/not_initialize.txtar asserts that lsp initialize must be
  called prior to any other commands (for this I had to update the
  server code accordingly)
- testdata/document_hover.txtar asserts the lsp document/hover command
- testdata/document_open.txtar assets the lsp document/open command

This is only the beginning as I have introduced many other tests in
gnols but pushing all scripts requires some other server changes and I'd
rather do that incrementally.